1. we need to switch an instance to a markRemoved rather than removing it and adding a new instance (the later plan breaks constraints on trovetroves) 2. remove code depended on removing things in an order which broke constraints on trovefiles and provides/requires -- we need to keep temp data to avoid that

+++ conary/repository/netrepos/trovestore.py
@@ -874,7 +874,7 @@ class TroveStore:
     def rollback(self):
         return self.db.rollback()
 
-    def removeTrove(self, name, version, flavor):
+    def _removeTrove(self, name, version, flavor, markOnly = False):
         cu = self.db.cursor()
         cu.execute("""
                 SELECT instanceId, itemId, Instances.versionId,
@@ -899,12 +899,8 @@ class TroveStore:
                    "WHERE itemId = ? AND versionId = ?", itemId, versionId)
         nodeId, branchId = cu.next()
 
-        cu.execute("DELETE FROM Provides WHERE instanceId = ?", instanceId)
-        cu.execute("DELETE FROM Requires WHERE instanceId = ?", instanceId)
-
         # remove all dependencies which are used only by this instanceId
         cu.execute("""
-        DELETE FROM Dependencies WHERE depId IN (
             SELECT AllDepsUsed.depId AS depId FROM
                 (SELECT depId FROM Provides WHERE instanceId = ?
                  UNION
@@ -929,9 +925,16 @@ class TroveStore:
                     AllDepsUsed.depId = DepsStillNeeded.depId
                 WHERE
                     DepsStillNeeded.depId IS NULL
-            )
         """, instanceId, instanceId, instanceId, instanceId, instanceId, 
              instanceId)
+        depsToRemove = [ x[0] for x in cu ]
+
+        cu.execute("DELETE FROM Provides WHERE instanceId = ?", instanceId)
+        cu.execute("DELETE FROM Requires WHERE instanceId = ?", instanceId)
+
+        if depsToRemove:
+            cu.execute("DELETE FROM Dependencies WHERE depId IN (%s)"
+                       % ",".join([ "%d" % x for x in depsToRemove ]))
 
         # Remove from TroveInfo
         cu.execute("DELETE FROM TroveInfo WHERE instanceId = ?", instanceId)
@@ -944,7 +947,7 @@ class TroveStore:
         # Now remove the files. Gather a list of sha1s of files to remove
         # from the filestore.
         cu.execute("""
-            SELECT sha1 FROM FileStreams WHERE streamId IN (
+            SELECT streamId, sha1 FROM FileStreams WHERE streamId IN (
                 SELECT TroveFiles.streamId FROM
                     TroveFiles LEFT OUTER JOIN
                         (SELECT FilesToKeep.streamId AS streamId
@@ -962,29 +965,22 @@ class TroveStore:
                         FilesToKeep.streamId is NULL
                 )
         """, instanceId, instanceId, instanceId)
-        filesToRemove = [ x[0] for x in cu ]
+        r = cu.fetchall()
+        filesToRemove = [ x[1] for x in r ]
+        streamIdsToRemove = [ x[0] for x in r ]
 
         cu.execute("DELETE FROM TroveFiles WHERE instanceId = ?", instanceId)
+        if streamIdsToRemove:
+            cu.execute("DELETE FROM FileStreams WHERE streamId IN (%s)"
+                       % ",".join([ "%d" % x for x in streamIdsToRemove ]))
 
-        cu.execute("""
-            DELETE FROM FileStreams WHERE streamId IN (
-                SELECT TroveFiles.streamId FROM
-                    TroveFiles LEFT OUTER JOIN
-                        (SELECT FilesToKeep.streamId AS streamId
-                            FROM TroveFiles AS FilesUsed
-                            JOIN TroveFiles AS FilesToKeep ON
-                                FilesUsed.streamId = FilesToKeep.streamId AND
-                                FilesToKeep.instanceId != ?
-                            WHERE
-                                FilesUsed.instanceId = ?
-                        ) AS FilesToKeep
-                    ON
-                        TroveFiles.streamId = FilesToKeep.streamId
-                    WHERE
-                        TroveFiles.instanceId = ? AND
-                        FilesToKeep.streamId is NULL
-                )
-        """, instanceId, instanceId, instanceId)
+        if markOnly:
+            # We don't actually remove anything here; we just mark the trove
+            # as removed instead
+            cu.execute("UPDATE Instances SET troveType=? WHERE instanceId=?",
+                       trove.TROVE_TYPE_REMOVED, instanceId)
+            self.versionOps.updateLatest(itemId, branchId, flavorId)
+            return filesToRemove
 
         cu.execute("DELETE FROM Instances WHERE instanceId = ?", instanceId)
 
@@ -1065,14 +1061,7 @@ class TroveStore:
         return filesToRemove
 
     def markTroveRemoved(self, name, version, flavor):
-        trv = trove.Trove(name, version, flavor, None,
-                          type = trove.TROVE_TYPE_REMOVED)
-
-        shas = self.removeTrove(name, version, flavor)
-        info = self.addTrove(trv)
-        self.addTroveDone(info)
-
-        return shas
+        return self._removeTrove(name, version, flavor, markOnly = True)
 
     def commit(self):
 	if self.needsCleanup:
